data model fix, some tests, jwt

+++ src/main/java/com/training/springbootbuyitem/entity/request/CreateItemRequestDto.java
@@ -8,6 +8,8 @@ import lombok.NoArgsConstructor;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Data
 @Builder
@@ -22,8 +24,8 @@ public class CreateItemRequestDto {
 	private String market;
 	@NotNull
 	@PositiveOrZero
-	private Integer stock;
+	private BigInteger stock;
 	@NotNull
 	@PositiveOrZero
-	private Double priceTag;
+	private BigDecimal priceTag;
 }
